Move pytest cli args to their own file
